# openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/posts/{postId}/comments": {
      "post": {
        "operationId": "PostCommentController_create",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "게시글 id",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "루트 댓글 작성. 답글은 POST /comments/:id/replies. 응답 = 단건 조회 표현(재조회)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "errorCode",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "errorCode": {
                      "type": "string",
                      "enum": [
                        "UNAUTHENTICATED"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "errorCode",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "errorCode": {
                      "type": "string",
                      "enum": [
                        "POST_NOT_FOUND"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "errorCode",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 422
                    },
                    "errorCode": {
                      "type": "string",
                      "enum": [
                        "INVALID_COMMENT_CONTENT"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "PostComment"
        ]
      },
      "get": {
        "operationId": "PostCommentController_list",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "게시글 id",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "이전 페이지 마지막 댓글 id. 생략 시 첫 페이지. 삭제된 id여도 동작(존재 검사 없음)",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지 크기 (1–50, 기본 20)",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "루트 댓글 목록 — 등록순(id ASC) keyset 페이지네이션. 삭제된 댓글은 답글이 남은 경우에만 deleted: true 플레이스홀더로 나온다. 답글은 replyCount만 주고 지연 로드. 인증 시 서로 차단한 관계의 댓글이 목록·replyCount에서 제외된다(비로그인이면 전체 노출)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentListDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "errorCode",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "errorCode": {
                      "type": "string",
                      "enum": [
                        "UNAUTHENTICATED"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "errorCode",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "errorCode": {
                      "type": "string",
                      "enum": [
                        "POST_NOT_FOUND"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "PostComment"
        ]
      }
    },
...
  },
  "info": {
    "title": "pulmung API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "description": "플레인텍스트 (HTML 아님 — 이스케이프는 렌더 몫). 1–2000자",
            "example": "저희 집 몬스테라도 그랬는데 물 주기를 늘리니 좋아졌어요"
          }
        },
        "required": [
          "content"
        ]
      },
      "CommentDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "parentId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              {
                "type": "null"
              }
            ],
            "description": "null = 루트 댓글, 값 = 이 답글이 속한 루트 댓글 id"
          },
          "content": {
            "type": "string",
            "example": "저희 집 몬스테라도 그랬는데 물 주기를 늘리니 좋아졌어요"
          },
          "author": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "nickname": {
                "type": "string",
                "example": "식집사"
              }
            },
            "required": [
              "id",
              "nickname"
            ],
            "additionalProperties": false
          },
          "mentionedUser": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "nickname": {
                    "type": "string",
                    "example": "식집사"
                  }
                },
                "required": [
                  "id",
                  "nickname"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ],
            "description": "답글이 지목한 유저(\"답글에 답글\"의 구조화 멘션) — 루트 댓글·일반 답글이면 null. 렌더 시 닉네임을 본문 앞에 @로 붙이는 건 클라 몫(서버는 텍스트에 섞지 않는다)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "id",
          "parentId",
          "content",
          "author",
          "mentionedUser",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "CommentListDto_Output": {
        "type": "object",
        "properties": {
          "comments": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "boolean",
                      "const": false
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "content": {
                      "type": "string",
                      "example": "저희 집 몬스테라도 그랬는데 물 주기를 늘리니 좋아졌어요"
                    },
                    "author": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "nickname": {
                          "type": "string",
                          "example": "식집사"
                        }
                      },
                      "required": [
                        "id",
                        "nickname"
                      ],
                      "additionalProperties": false
                    },
                    "replyCount": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991,
                      "description": "이 루트에 달린 답글 수 — \"답글 N개 보기\" 렌더용. 상세는 GET /comments/:id/replies"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    }
                  },
                  "required": [
                    "deleted",
                    "id",
                    "content",
                    "author",
                    "replyCount",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                },
                {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "boolean",
                      "const": true,
                      "description": "답글이 남아 있어 자리만 보존된 삭제 댓글 — 클라가 \"삭제된 댓글입니다\"로 렌더"
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "replyCount": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991,
                      "description": "이 루트에 달린 답글 수 — \"답글 N개 보기\" 렌더용. 상세는 GET /comments/:id/replies"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    }
                  },
                  "required": [
                    "deleted",
                    "id",
                    "replyCount",
                    "createdAt"
                  ],
                  "additionalProperties": false
                }
              ]
            },
            "description": "루트 댓글만(등록순) — 답글은 GET /comments/:id/replies로 지연 로드"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              {
                "type": "null"
              }
            ],
            "description": "다음 페이지 cursor — null이면 마지막 페이지"
          }
        },
        "required": [
          "comments",
          "nextCursor"
        ],
        "additionalProperties": false
      },
...
    }
  }
}